+++ src/designer_plugin/d3sdk/__init__.py
@@ -5,7 +5,7 @@ Copyright (c) 2025 Disguise Technologies ltd
 
 from .client import D3PluginClient
 from .function import (
-    add_packages_in_current_file,
+    PackageInfo,
     d3function,
     d3pythonscript,
     get_all_d3functions,
@@ -18,9 +18,9 @@ __all__: list[str] = [
     "D3AsyncSession",
     "D3PluginClient",
     "D3Session",
+    "PackageInfo",
     "d3pythonscript",
     "d3function",
-    "add_packages_in_current_file",
     "get_register_payload",
     "get_all_d3functions",
     "get_all_modules",